rename: change MobileMenu.js to Sidebar.js
